backend/requirements.txt
Flask==1.1.2
pyjwt==1.7.1
flask-cors==3.0.8
google-auth==1.15.0
requests==2.23.0
requests-toolbelt==0.9.1
google-python-cloud-debugger